backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# MCP Server
mcp>=1.0.0

# Image Upscaling
transformers>=4.35.0
torch>=2.1.0
Pillow>=10.0.0
scikit-image>=0.22.0
